+++ src/views/domain-workflows/domain-workflows-header/domain-workflows-header.types.ts
@@ -2,3 +2,8 @@ import { type ListWorkflowsRequestQueryParams } from '@/route-handlers/list-work
 
 export type DomainWorkflowsHeaderInputType =
   ListWorkflowsRequestQueryParams['inputType'];
+
+export type Props = {
+  domain: string;
+  cluster: string;
+};
