+++ src/views/domain-workflows/domain-workflows-query-input/domain-workflows-query-input.types.ts
@@ -1,4 +1,5 @@
 export type Props = {
   value: string;
   setValue: (v: string | undefined) => void;
+  refetchQuery: () => void;
 };
