--- src/views/domain-workflows/domain-workflows-table-search/domain-workflows-table-search.styles.ts
@@ -1,10 +1,0 @@
-import { styled as createStyled } from 'baseui';
-
-export const styled = {
-  ErrorPanelContainer: createStyled('div', ({ $theme }) => ({
-    padding: `${$theme.sizing.scale1200} 0px`,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  })),
-};
