+++ src/views/domain-workflows/domain-workflows-table/domain-workflows-table.types.ts
@@ -1,5 +1,4 @@
 export type Props = {
   domain: string;
   cluster: string;
-  // TODO @adhitya.mamallan - when working on Archival, add a flag here
 };
